requirements.txt
Flask
requests
pycryptodome
